+++ sdk/redis/arm-rediscache/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release History
     
+## 8.1.0 (2024-04-16)
+    
+**Features**
+
+  - Interface RedisCommonPropertiesRedisConfiguration has a new optional parameter notifyKeyspaceEvents
+    
+    
 ## 8.0.0 (2023-09-11)
     
 **Features**
@@ -156,4 +163,4 @@ To understand the detail of the change, please refer to [Changelog](https://aka.
 
 To migrate the existing applications to the latest version, please refer to [Migration Guide](https://aka.ms/js-track2-migration-guide).
 
-To learn more, please refer to our documentation [Quick Start](https://aka.ms/azsdk/js/mgmt/quickstart ).
+To learn more, please refer to our documentation [Quick Start](https://aka.ms/js-track2-quickstart).